+++ elyra/metadata/handlers.py
@@ -1,5 +1,5 @@
 #
-# Copyright 2018-2020 IBM Corporation
+# Copyright 2018-2020 Elyra Authors
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -15,9 +15,9 @@
 #
 
 from jsonschema import ValidationError
+from jupyter_server.base.handlers import APIHandler
+from jupyter_server.utils import url_unescape, url_path_join
 from tornado import web
-from notebook.base.handlers import APIHandler
-from notebook.utils import url_unescape, url_path_join
 
 from .error import MetadataNotFoundError, MetadataExistsError, SchemaNotFoundError
 from .metadata import Metadata