+++ elyra/metadata/tests/test_metadata_app.py
@@ -1,5 +1,5 @@
 #
-# Copyright 2018-2020 IBM Corporation
+# Copyright 2018-2020 Elyra Authors
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -22,7 +22,8 @@ import shutil
 from tempfile import mkdtemp
 from elyra.metadata import Metadata, MetadataManager, METADATA_TEST_NAMESPACE
 from .test_utils import PropertyTester, create_json_file, valid_metadata_json, \
-    another_metadata_json, invalid_metadata_json, invalid_no_display_name_json
+    another_metadata_json, invalid_metadata_json, invalid_no_display_name_json, \
+    invalid_schema_name_json
 
 os.environ["METADATA_TESTING"] = "1"  # Enable metadata-tests namespace
 
@@ -178,7 +179,6 @@ def test_install_and_replace(script_runner, mock_data_dir):
                             '--name=test-metadata_42_valid-name', '--display_name=display_name',
                             '--required_test=required_value')
     assert ret.success is False
-    assert ret.stdout == ''
     assert "An instance named 'test-metadata_42_valid-name' already exists in the metadata-tests " \
            "namespace" in ret.stderr
 
@@ -263,13 +263,14 @@ def test_list_instances(script_runner, mock_data_dir):
     metadata_dir = os.path.join(mock_data_dir, 'metadata', METADATA_TEST_NAMESPACE)
     create_json_file(metadata_dir, 'invalid.json', invalid_metadata_json)
     create_json_file(metadata_dir, 'no_display_name.json', invalid_no_display_name_json)
+    create_json_file(metadata_dir, 'invalid_schema_name.json', invalid_schema_name_json)
 
     ret = script_runner.run('elyra-metadata', 'list', METADATA_TEST_NAMESPACE)
     assert ret.success
     lines = ret.stdout.split('\n')
-    assert len(lines) == 9  # always 5 more than the actual runtime count
+    assert len(lines) == 10  # always 5 more than the actual runtime count
     assert lines[0] == "Available metadata instances for {} (includes invalid):".format(METADATA_TEST_NAMESPACE)
-    line_elements = [line.split() for line in lines[4:8]]
+    line_elements = [line.split() for line in lines[4:9]]
     assert line_elements[0][1] == "another"
     assert line_elements[1][1] == "invalid"
     assert line_elements[1][3] == "**INVALID**"
@@ -277,6 +278,8 @@ def test_list_instances(script_runner, mock_data_dir):
     assert line_elements[2][3] == "**INVALID**"
     assert line_elements[2][4] == "(ValidationError)"
     assert line_elements[3][1] == "valid"
+    assert line_elements[4][3] == "**INVALID**"
+    assert line_elements[4][4] == "(SchemaNotFoundError)"
 
     ret = script_runner.run('elyra-metadata', 'list', METADATA_TEST_NAMESPACE, '--valid-only')
     assert ret.success