+++ docs/source/getting_started/changelog.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 1.4.2 (12/04/2020)
+
+- Enable Elyra image to auto-start JupyterLab
+- Update to JupyterHub SingleUser 0.10.6 docker image [#1135](https://github.com/elyra-ai/elyra/pull/1135)
+- Remove workdir setting from Elyra docker image [#1135](https://github.com/elyra-ai/elyra/pull/1135)
+- Build both docker.io and quay.io tagged images [#1135](https://github.com/elyra-ai/elyra/pull/1135)
+- Update Jupyter Client minimum version to 6.1.7  [#1099](https://github.com/elyra-ai/elyra/pull/1099)
+- Properly handle references to missing schema files [#1096](https://github.com/elyra-ai/elyra/pull/1096)
+- Support namespace configuration when using dex with kfp [#1081](https://github.com/elyra-ai/elyra/pull/1081)
+- Update canvas version [#1082](https://github.com/elyra-ai/elyra/pull/1082)
+
 ## 1.4.1 (11/14/2020)
 
 - Use JupyterLab SessionManager to execute Python scripts [#1071](https://github.com/elyra-ai/elyra/pull/1071)